Add LevelController and relevant view
Implemented a new LevelController and corresponding view to retrieve and display user levels data. The LevelController includes an index method which retrieves data from the 'm_level' table and passes this data to the view. This view then displays the id, kode, and nama of each level in a HTML table. The route has also been added.